requirements.txt
python-telegram-bot==20.3
requests==2.31.0
python-dotenv==1.0.0
APScheduler==3.10.4
aiohttp==3.9.1
Pillow==10.1.0
pytz==2024.1

